Conflicts:
	imap/http_caldav.c

+++ imap/mailbox.c
@@ -109,6 +109,9 @@
 #include "xstrlcat.h"
 #include "xstats.h"
 
+#include "objectstore.h"
+#include "objectstore_db.h"
+
 /* generated headers are not necessarily in current directory */
 #include "imap/imap_err.h"
 
@@ -164,6 +167,9 @@ static struct mailboxlist *create_listitem(const char *name)
     /* ensure we never print insane times */
     gettimeofday(&item->m.starttime, 0);
 
+    if (config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED))
+        keep_user_message_db_open (1);
+
     return item;
 }
 
@@ -186,11 +192,16 @@ static void remove_listitem(struct mailboxlist *remitem)
 
     for (item = open_mailboxes; item; item = item->next) {
         if (item == remitem) {
+
             if (previtem)
                 previtem->next = item->next;
             else
                 open_mailboxes = item->next;
             free(item);
+
+            if (!open_mailboxes && config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED))  // time to force clese the database
+                 keep_user_message_db_open (0);
+
             return;
         }
         previtem = item;
@@ -247,7 +258,7 @@ static const char *mailbox_archive_fname(struct mailbox *mailbox, uint32_t uid)
 EXPORTED const char *mailbox_record_fname(struct mailbox *mailbox,
                                           const struct index_record *record)
 {
-    if (record->system_flags & FLAG_ARCHIVED)
+    if (!config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED) && record->system_flags & FLAG_ARCHIVED)
         return mailbox_archive_fname(mailbox, record->uid);
     else
         return mailbox_spool_fname(mailbox, record->uid);
@@ -788,14 +799,11 @@ EXPORTED void mailbox_make_uniqueid(struct mailbox *mailbox)
     mailbox_set_uniqueid(mailbox, makeuuid());
 }
 
-EXPORTED int mailbox_map_record(struct mailbox *mailbox, const struct index_record *record, struct buf *buf)
+
+static int _map_local_record(const struct mailbox *mailbox, const char *fname, struct buf *buf)
 {
-    const char *fname;
     struct stat sbuf;
     int msgfd;
-
-    xstats_inc(MESSAGE_MAP);
-    fname = mailbox_record_fname(mailbox, record);
 
     msgfd = open(fname, O_RDONLY, 0666);
     if (msgfd == -1) return errno;
@@ -810,6 +818,25 @@ EXPORTED int mailbox_map_record(struct mailbox *mailbox, const struct index_reco
 
     return 0;
 }
+
+EXPORTED int mailbox_map_record(struct mailbox *mailbox, const struct index_record *record, struct buf *buf)
+{
+    xstats_inc(MESSAGE_MAP);
+
+    const char *fname = mailbox_record_fname(mailbox, record);
+    int r = _map_local_record(mailbox, fname, buf);
+    if (!r) return 0;
+
+    if (config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED)){
+        r = objectstore_get(mailbox, record, fname);
+        if (r) return r;
+        r = _map_local_record(mailbox, fname, buf);
+        remove(fname);
+    }
+
+    return r;
+}
+
 
 static void mailbox_release_resources(struct mailbox *mailbox)
 {
@@ -1105,7 +1132,7 @@ EXPORTED void mailbox_close(struct mailbox **mailboxptr)
         if (!r) {
             /* finish cleaning up */
             if (mailbox->i.options & OPT_MAILBOX_DELETED)
-                mailbox_delete_cleanup(mailbox->part, mailbox->name, mailbox->uniqueid);
+                mailbox_delete_cleanup(mailbox, mailbox->part, mailbox->name, mailbox->uniqueid);
             else if (mailbox->i.options & OPT_MAILBOX_NEEDS_REPACK)
                 mailbox_index_repack(mailbox, mailbox->i.minor_version);
             else if (mailbox->i.options & OPT_MAILBOX_NEEDS_UNLINK)
@@ -3687,8 +3714,9 @@ EXPORTED int mailbox_append_index_record(struct mailbox *mailbox,
     if (!(record->system_flags & FLAG_UNLINKED)) {
         /* make the file timestamp correct */
         settime.actime = settime.modtime = record->internaldate;
-        if (utime(mailbox_record_fname(mailbox, record), &settime) == -1)
-            return IMAP_IOERROR;
+        if (!(config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED) && record->system_flags & FLAG_ARCHIVED ))  // mabe there is no file in directory.
+            if (utime(mailbox_record_fname(mailbox, record), &settime) == -1)
+                return IMAP_IOERROR;
 
         /* write the cache record before buffering the message, it
          * will set the cache_offset field. */
@@ -3718,6 +3746,7 @@ static void mailbox_record_cleanup(struct mailbox *mailbox,
 {
     const char *spoolfname = mailbox_spool_fname(mailbox, record->uid);
     const char *archivefname = mailbox_archive_fname(mailbox, record->uid);
+    int object_storage_enabled = config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED) ;
     int r;
 
     if (record->system_flags & FLAG_UNLINKED) {
@@ -3731,12 +3760,17 @@ static void mailbox_record_cleanup(struct mailbox *mailbox,
                        record->uid, record->system_flags);
         }
 
-        if (unlink(archivefname) == 0) {
-            if (config_auditlog)
-                syslog(LOG_NOTICE, "auditlog: unlinkarchive sessionid=<%s> "
-                       "mailbox=<%s> uniqueid=<%s> uid=<%u> sysflags=<%u>",
-                       session_id(), mailbox->name, mailbox->uniqueid,
-                       record->uid, record->system_flags);
+        if (object_storage_enabled) {
+            objectstore_delete(mailbox, record);
+        }
+        else if (strcmp(spoolfname, archivefname)) {
+            if (unlink(archivefname) == 0) {
+                if (config_auditlog)
+                    syslog(LOG_NOTICE, "auditlog: unlinkarchive sessionid=<%s> "
+                            "mailbox=<%s> uniqueid=<%s> uid=<%u> sysflags=<%u>",
+                            session_id(), mailbox->name, mailbox->uniqueid,
+                            record->uid, record->system_flags);
+            }
         }
 
         r = mailbox_get_annotate_state(mailbox, record->uid, NULL);
@@ -3754,19 +3788,19 @@ static void mailbox_record_cleanup(struct mailbox *mailbox,
         }
     }
 
-    /* nothing to cleanup if it's the same file! */
-    if (!strcmp(spoolfname, archivefname))
-        return;
-
-    /* otherwise check for archived/nonarchived */
-    else if (record->system_flags & FLAG_ARCHIVED) {
+    if (record->system_flags & FLAG_ARCHIVED) {
         /* XXX - stat to make sure the other file exists first? - we mostly
-         * trust that we didn't do stupid things everywhere else, so maybe not */
+         *  trust that we didn't do stupid things everywhere else, so maybe not */
         unlink(spoolfname);
     }
-
     else {
-        unlink(archivefname);
+        if (object_storage_enabled) {
+            objectstore_delete(mailbox, record);
+        }
+        else if (strcmp(spoolfname, archivefname)) {
+            unlink(archivefname);
+        }
+        /* else: nothing to cleanup if it's the same file! */
     }
 }
 
@@ -4197,6 +4231,15 @@ static unsigned expungedeleted(struct mailbox *mailbox __attribute__((unused)),
     return 0;
 }
 
+//  this is used to get message back from ObjectStore
+EXPORTED unsigned mailbox_should_unarchive(struct mailbox *mailbox __attribute__((unused)),
+                                           const struct index_record *record __attribute__((unused)),
+                                           void *rock __attribute__((unused)))
+{
+    return 0;
+}
+
+
 EXPORTED unsigned mailbox_should_archive(struct mailbox *mailbox,
                                          const struct index_record *record,
                                          void *rock)
@@ -4241,7 +4284,7 @@ EXPORTED unsigned mailbox_should_archive(struct mailbox *mailbox,
         return 0;
 
     /* archive all other old messages */
-    if (record->internaldate < cutoff)
+    if (record->internaldate <= cutoff)
         return 1;
 
     /* and don't archive anything else! */
@@ -4255,7 +4298,8 @@ EXPORTED unsigned mailbox_should_archive(struct mailbox *mailbox,
  * should be in the spool - if 1, the message should be in the archive.
  */
 EXPORTED void mailbox_archive(struct mailbox *mailbox,
-                              mailbox_decideproc_t *decideproc, void *deciderock)
+                              mailbox_decideproc_t *decideproc, void *deciderock, unsigned flags)
+
 {
     int r;
     int dirtycache = 0;
@@ -4266,13 +4310,15 @@ EXPORTED void mailbox_archive(struct mailbox *mailbox,
     char *spoolcache = xstrdup(mailbox_meta_fname(mailbox, META_CACHE));
     char *archivecache = xstrdup(mailbox_meta_fname(mailbox, META_ARCHIVECACHE));
     int differentcache = strcmp(spoolcache, archivecache);
+    int object_storage_enabled = config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED) ;
     free(spoolcache);
     free(archivecache);
 
     assert(mailbox_index_islocked(mailbox, 1));
     if (!decideproc) decideproc = &mailbox_should_archive;
 
-    struct mailbox_iter *iter = mailbox_iter_init(mailbox, 0, ITER_SKIP_EXPUNGED);
+    struct mailbox_iter *iter = mailbox_iter_init(mailbox, 0, flags);
+
     while ((record = mailbox_iter_step(iter))) {
         const char *action = NULL;
         if (decideproc(mailbox, record, deciderock)) {
@@ -4289,7 +4335,12 @@ EXPORTED void mailbox_archive(struct mailbox *mailbox,
                        mailbox->name, copyrecord.uid, error_message(r));
                 continue;
             }
-
+            if (object_storage_enabled){
+                /* upload on the blob store */
+                r = objectstore_put(mailbox, &copyrecord, srcname);
+                if (r)
+                    continue;
+            }
             copyrecord.system_flags |= FLAG_ARCHIVED | FLAG_NEEDS_CLEANUP;
             action = "archive";
         }
@@ -4307,19 +4358,39 @@ EXPORTED void mailbox_archive(struct mailbox *mailbox,
                        mailbox->name, copyrecord.uid, error_message(r));
                 continue;
             }
-
             copyrecord.system_flags &= ~FLAG_ARCHIVED;
             copyrecord.system_flags |= FLAG_NEEDS_CLEANUP;
             action = "unarchive";
         }
 
-        /* got a file to copy! */
-        if (strcmp(srcname, destname)) {
-            r = cyrus_copyfile(srcname, destname, COPYFILE_MKDIR);
-            if (r) {
-                syslog(LOG_ERR, "IOERROR archive %s %u failed to copyfile (%s => %s): %s",
-                       mailbox->name, copyrecord.uid, srcname, destname, error_message(r));
-                continue;
+        if (object_storage_enabled) {
+            /* get or fetch */
+            if (copyrecord.system_flags & FLAG_ARCHIVED) {
+                r = objectstore_put(mailbox, &copyrecord, srcname);
+                if (r) {
+                    syslog(LOG_ERR, "IOERROR archive %s %u failed to objectstorage put file (%s): %s",
+                           mailbox->name, copyrecord.uid, srcname, error_message(r));
+                    continue;
+                }
+            }
+            else {
+                r = objectstore_get(mailbox, &copyrecord, destname);
+                if (r) {
+                    syslog(LOG_ERR, "IOERROR archive %s %u failed to objectstorage get file (%s): %s",
+                           mailbox->name, copyrecord.uid, destname, error_message(r));
+                    continue;
+                }
+            }
+        }
+        else {
+            /* got a file to copy! */
+            if (strcmp(srcname, destname)) {
+                r = cyrus_copyfile(srcname, destname, COPYFILE_MKDIR);
+                if (r) {
+                    syslog(LOG_ERR, "IOERROR archive %s %u failed to copyfile (%s => %s): %s",
+                            mailbox->name, copyrecord.uid, srcname, destname, error_message(r));
+                    continue;
+                }
             }
         }
 
@@ -4962,7 +5033,7 @@ static struct meta_file meta_files[] = {
  * try to create a mailbox while the delete is underway.
  * VERY tight race condition exists right now... */
 /* we need an exclusive namelock for this */
-HIDDEN int mailbox_delete_cleanup(const char *part, const char *name, const char *uniqueid)
+HIDDEN int mailbox_delete_cleanup(struct mailbox *mailbox, const char *part, const char *name, const char *uniqueid)
 {
     strarray_t paths = STRARRAY_INITIALIZER;
     int i;
@@ -4973,8 +5044,14 @@ HIDDEN int mailbox_delete_cleanup(const char *part, const char *name, const char
     char *ntail;
     char *p;
 
+    int object_storage_enabled = config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED);
+
     strncpy(nbuf, name, sizeof(nbuf));
     ntail = nbuf + strlen(nbuf);
+
+    if (mailbox && object_storage_enabled){
+        mailbox_archive(mailbox, mailbox_should_unarchive, NULL, 0);
+    }
 
     /* XXX - double XXX - this is a really ugly function.  It should be
      * using mboxname_parts and walking back up the 'boxes' list */
@@ -5047,6 +5124,7 @@ EXPORTED int mailbox_copy_files(struct mailbox *mailbox, const char *newpart,
     struct meta_file *mf;
     const struct index_record *record;
     int r = 0;
+    int object_storage_enabled = config_getswitch(IMAPOPT_OBJECT_STORAGE_ENABLED);
 
     /* Copy over meta files */
     for (mf = meta_files; mf->metaflag; mf++) {
@@ -5069,15 +5147,26 @@ EXPORTED int mailbox_copy_files(struct mailbox *mailbox, const char *newpart,
     while ((record = mailbox_iter_step(iter))) {
         xstrncpy(oldbuf, mailbox_record_fname(mailbox, record),
                 MAX_MAILBOX_PATH);
-        if (record->system_flags & FLAG_ARCHIVED)
+        if (!object_storage_enabled && record->system_flags & FLAG_ARCHIVED)
             xstrncpy(newbuf, mboxname_archivepath(newpart, newname, newuniqueid, record->uid),
                     MAX_MAILBOX_PATH);
         else
             xstrncpy(newbuf, mboxname_datapath(newpart, newname, newuniqueid, record->uid),
                     MAX_MAILBOX_PATH);
 
-        r = mailbox_copyfile(oldbuf, newbuf, 0);
+        if (!(object_storage_enabled && record->system_flags & FLAG_ARCHIVED ))    // if object storage do not move file
+           r = mailbox_copyfile(oldbuf, newbuf, 0);
+
         if (r) break;
+
+        if (object_storage_enabled) {
+            static struct mailbox new_mailbox;
+            memset(&new_mailbox, 0, sizeof(struct mailbox));
+            new_mailbox.name = (char*) newname;
+            new_mailbox.part = (char*) config_defpartition ;
+            r = objectstore_put(&new_mailbox, record, newbuf);   // put should just add to refcount.
+            if (r) break;
+        }
     }
     mailbox_iter_done(&iter);
 
@@ -5222,7 +5311,7 @@ fail:
     /* first unlock so we don't need to write anything new down */
     mailbox_unlock_index(newmailbox, NULL);
     /* then remove all the files */
-    mailbox_delete_cleanup(newmailbox->part, newmailbox->name, newmailbox->uniqueid);
+    mailbox_delete_cleanup(NULL, newmailbox->part, newmailbox->name, newmailbox->uniqueid);
     /* and finally, abort */
     mailbox_abort(newmailbox);
     mailbox_close(&newmailbox);
@@ -5233,6 +5322,7 @@ fail:
 
 EXPORTED int mailbox_rename_cleanup(struct mailbox **mailboxptr, int isinbox)
 {
+
     int r = 0;
     struct mailbox *oldmailbox = *mailboxptr;
     char *name = xstrdup(oldmailbox->name);
